tests/submission/W3C/domhtml/HTMLAnchorElement12.html

<!DOCTYPE html>
<html><head>
<meta charset="utf-8">
<title>NIST DOM HTML Test - Anchor</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../common/domtests-common.js"></script>
<link href="http://www.w3.org/TR/html5/the-a-element.html#dom-a-type" rel="help">
<link rel="author" title="NIST">
</head>
<body>
<p>
<a href="/support/images/yellow-rectangle.png" type="image/png">link</a>
</p>


<script>
var domtest = async_test("The type attribute contains the advisory content model.");

on_event(window, "load", domtest.step_func(function (dom_event) {
    var nodeList;
      var testNode;
      var vtype;
      var doc;
      doc = window.document;
      nodeList = doc.getElementsByTagName("a");
      assert_equals(nodeList.length,1);
testNode = nodeList.item(0);
      vtype = testNode.type;

      assert_equals(vtype,"image/png");
       
    domtest.done();
}));

</script>
<p>The type attribute contains the advisory content model. 

    Retrieve the type attribute and examine its value.</p>
<div id="log"></div>
</body></html>
